rename follower mod to chain_state
