+++ catalyst-gateway/bin/src/event_db/cardano/chain_state/insert_update_state.sql
