+++ catalyst-gateway/bin/src/event_db/cardano/chain_state/select_slot_info_by_datetime.sql.hbs
