+++ catalyst-gateway/bin/src/event_db/cardano/chain_state/select_update_state.sql
