+++ catalyst-gateway/bin/src/event_db/cardano/mod.rs
@@ -1,6 +1,6 @@
 //! Cardano db functionality
 
 pub(crate) mod config;
-pub(crate) mod follower;
+pub(crate) mod chain_state;
 pub(crate) mod utxo;
 pub(crate) mod voter_registration;
