+++ catalyst-gateway/bin/src/service/api/cardano/mod.rs
@@ -9,7 +9,7 @@ use poem_openapi::{
 };
 
 use crate::{
-    event_db::cardano::follower::{DateTime, SlotNumber},
+    event_db::cardano::chain_state::{DateTime, SlotNumber},
     service::{
         common::{
             objects::cardano::{network::Network, stake_address::StakeAddress},
