+++ catalyst-gateway/bin/src/service/common/objects/cardano/stake_info.rs
@@ -2,7 +2,7 @@
 
 use poem_openapi::{types::Example, Object};
 
-use crate::event_db::cardano::{follower::SlotNumber, utxo::StakeAmount};
+use crate::event_db::cardano::{chain_state::SlotNumber, utxo::StakeAmount};
 
 /// User's cardano stake info.
 #[derive(Object)]
